# api-docs.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/prompts": {
      "get": {
        "operationId": "PromptController_listPrompts",
        "parameters": [
          {
            "name": "X-Locale",
            "in": "header",
            "description": "Locale header. Defaults to en.",
            "required": false,
            "schema": { "type": "string" }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": { "default": 1, "type": "number" }
          },
          {
            "name": "page_size",
            "required": false,
            "in": "query",
            "schema": { "default": 20, "type": "number" }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Search prompt name or description",
            "schema": { "type": "string" }
          },
          {
            "name": "media_type",
            "required": false,
            "in": "query",
            "description": "Prompt media type",
            "schema": { "type": "string" }
          },
          {
            "name": "category",
            "required": false,
            "in": "query",
            "description": "Filter by category code",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/PaginatedResponse" },
                    {
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/PromptResponse"
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [{ "user-token": [] }],
        "summary": "List prompts",
        "tags": ["Prompt"]
      }
    },
...
  },
  "info": {
    "title": "GaryTu AI API",
    "description": "",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "user-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "admin-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "PaginationMeta": {
        "type": "object",
        "properties": {
          "page": { "type": "number" },
          "page_size": { "type": "number" },
          "total": { "type": "number" },
          "total_pages": { "type": "number" }
        },
        "required": ["page", "page_size", "total", "total_pages"]
      },
      "PaginatedResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "object",
            "description": "Code, 0 for success, others for errors"
          },
          "message": {
            "type": "string",
            "description": "Message for developers"
          },
          "data": {
            "type": "object",
            "description": "Response data",
            "nullable": true
          },
          "timestamp": { "type": "string", "description": "ISO8601 timestamp" },
          "errors": {
            "description": "Optional errors array",
            "type": "array",
            "items": { "type": "object" }
          },
          "extra": { "type": "object", "description": "Extra info" },
          "meta": { "$ref": "#/components/schemas/PaginationMeta" }
        },
        "required": ["code", "message", "timestamp", "meta"]
      },
      "PromptCategoryResponse": {
        "type": "object",
        "properties": {
          "code": { "type": "string" },
          "name": { "type": "string" }
        },
        "required": ["code", "name"]
      },
      "PromptFileResponse": {
        "type": "object",
        "properties": {
          "uuid": { "type": "string" },
          "category": { "type": "string" },
          "file_type": { "type": "string" },
          "position": { "type": "number" },
          "url": { "type": "string" },
          "thumbnail_url": { "type": "object", "nullable": true },
          "created_at": { "type": "object", "nullable": true }
        },
        "required": ["uuid", "category", "file_type", "position", "url"]
      },
      "PromptLabelResponse": {
        "type": "object",
        "properties": {
          "code": { "type": "string" },
          "name": { "type": "string" }
        },
        "required": ["code", "name"]
      },
      "PromptUserStateResponse": {
        "type": "object",
        "properties": {
          "is_favorite": { "type": "boolean" },
          "purchased": { "type": "boolean" }
        },
        "required": ["is_favorite", "purchased"]
      },
      "PromptResponse": {
        "type": "object",
        "properties": {
          "uuid": { "type": "string" },
          "name": { "type": "string" },
          "description": { "type": "object", "nullable": true },
          "price": { "type": "number" },
          "enabled": { "type": "boolean" },
          "bonus_credit": { "type": "number" },
          "media_type": {
            "type": "string",
            "enum": ["IMAGE", "VIDEO"],
            "nullable": true
          },
          "category": {
            "nullable": true,
            "allOf": [{ "$ref": "#/components/schemas/PromptCategoryResponse" }]
          },
          "created_at": { "type": "object", "nullable": true },
          "cover": {
            "nullable": true,
            "allOf": [{ "$ref": "#/components/schemas/PromptFileResponse" }]
          },
          "pdf": {
            "nullable": true,
            "allOf": [{ "$ref": "#/components/schemas/PromptFileResponse" }]
          },
          "zip": {
            "nullable": true,
            "allOf": [{ "$ref": "#/components/schemas/PromptFileResponse" }]
          },
          "files": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/PromptFileResponse" }
          },
          "labels": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/PromptLabelResponse" }
          },
          "user_state": {
            "$ref": "#/components/schemas/PromptUserStateResponse"
          }
        },
        "required": [
          "uuid",
          "name",
          "price",
          "enabled",
          "bonus_credit",
          "files",
          "labels",
          "user_state"
        ]
      },
...
    }
  }
}